feat: preluare comenzi
