+++ src/App.js
@@ -52,155 +52,156 @@ const App = () => {
 
   return (
 
-      <div>
-        <Grid container spacing={2} justifyContent="center" padding={2} >
-          <Grid item padding={3}>
-            <Logo />
-          </Grid>
-          {!isOrderSent ?
-            <>
-              <Grid item size={12}>
-                <Typography variant="h4" align="center" fontWeight="bold">
-                  Formular comandă
-                </Typography>
-              </Grid>
-              <Grid item size={{ xs: 12, sm: 8, md: 6, lg: 7, xl: 4 }}>
-                <Formik
-                  initialValues={{
-                    name: '',
-                    phone: '',
-                    email: '',
-                    products: [],
-                    weight: '',
-                    id: '',
-                    consent: false,
-                    pickUpDate: '',
-                    pickUpTime: ''
-                  }}
-                  validationSchema={Yup.object({
-                    name: Yup.string().required('Este necesară completarea acestui câmp'),
-                    phone: Yup.string().required('Este necesară completarea acestui câmp'),
-                    email: Yup.string().email('Adresă de e-mail invalidă'),
-                    consent: Yup.boolean().required(),
-                    pickUpDate: Yup.string().required('Este necesar sa selectati o data de ridicare a comenzii dvs.'),
-                    pickUpTime: Yup.string().required('Este necesar sa selectati un interval de ridicare a comenzii dvs.'),
-                    products: Yup.array().of(
-                      Yup.object().shape({
-                        quantity: Yup.number().positive().required('Required'),
-                        id: Yup.number().positive().required('Required'),
-                      })
-                    ).min(1, 'The error message if length === 0 | 1'),
-                  })}
-                  onSubmit={(values, { setSubmitting }) => {
-                    const formData = new FormData();
-                    formData.append('name', values.name)
-                    formData.append('phone', values.phone)
-                    formData.append('email', values.email)
-                    formData.append('products', JSON.stringify(values.products));
-                    formData.append('pickUpDate', values.pickUpDate);
-                    formData.append('pickUpTime', values.pickUpTime);
-                    axios
-                      .post(
-                        'https://script.google.com/macros/s/AKfycbx_bdwXRlqbVmbLIy90U5HJG_KwTEjcreTsGJdEC7Bt95bZHzem1K4rFgPwRZES4OYz/exec',
-                        formData
-                      )
-                      .then((response) => {
-                        setIsOrderSent(true)
-                      })
-                      .catch((error) => {
-                        console.error('Error submitting form:', error);
-                        alert('There was an error submitting the form.');
-                      })
-                      .finally(() => {
-                        setSubmitting(false);
-                      });
-                  }}
-                >
-                  {({
-                    values,
-                    errors,
-                    touched,
-                    handleChange,
-                    handleBlur,
-                    isSubmitting,
-                    setFieldValue,
-                    isValid
-                  }) => (
-                    <Form>
-                      <Grid container alignContent={"center"} justifyContent={"center"}>
-                        <Grid item size={{ xs: 12, sm: 12, md: 12, lg: 8 }}>
-                          <TextField
-                            fullWidth
-                            id="name"
-                            name="name"
-                            label="Nume și prenume"
-                            value={values.name}
-                            onChange={handleChange}
-                            onBlur={handleBlur}
-                            error={touched.name && Boolean(errors.name)}
-                            helperText={touched.name && errors.name}
-                            margin="normal"
-                            disabled={isSubmitting}
-                            required
-                          />
-                          <TextField
-                            fullWidth
-                            id="phone"
-                            name="phone"
-                            label="Număr telefon"
-                            value={values.phone}
-                            onChange={handleChange}
-                            onBlur={handleBlur}
-                            error={touched.phone && Boolean(errors.phone)}
-                            helperText={touched.phone && errors.phone}
-                            margin="normal"
-                            disabled={isSubmitting}
-                            required
-                          />
-                          <TextField
-                            select
-                            fullWidth
-                            id="pickupDate"
-                            name="pickupDate"
-                            label="Data ridicare"
-                            value={values.pickupDate}
-                            onChange={(e) => {
-                              const selectedDate = e.target.value;
-                              setFieldValue('pickUpDate', selectedDate);
-                              setPickupIntervals(
-                                selectedDate === "24" ? intervalOptions.restricted : intervalOptions.default
-                              );
-                            }}
-                            margin="normal"
-                            required
-                            disabled={isSubmitting}
-                          >
-                            {dateOptions.map((date) => (
-                              <MenuItem key={date} value={date}>
-                                {`${date} Decembrie`}
-                              </MenuItem>
-                            ))}
-                          </TextField>
-                          <TextField
-                            select
-                            fullWidth
-                            id="pickupInterval"
-                            name="pickupInterval"
-                            label="Interval ridicare"
-                            value={values.pickupInterval}
-                            onChange={(e) => setFieldValue('pickUpTime', e.target.value)}
-                            margin="normal"
-                            required
-                            disabled={!values.pickUpDate || isSubmitting}
-                          >
-                            {pickupIntervals.map((interval, index) => (
-                              <MenuItem key={index} value={interval}>
-                                {interval}
-                              </MenuItem>
-                            ))}
-                          </TextField>
+    <div>
+      <Grid container spacing={2} justifyContent="center" padding={2} >
+        <Grid item padding={3}>
+          <Logo />
+        </Grid>
+        <Formik
+          initialValues={{
+            name: '',
+            phone: '',
+            email: '',
+            products: [],
+            weight: '',
+            id: '',
+            consent: false,
+            pickUpDate: '30',
+            pickUpTime: '09-13'
+          }}
+          validationSchema={Yup.object({
+            name: Yup.string().required('Este necesară completarea acestui câmp'),
+            phone: Yup.string().required('Este necesară completarea acestui câmp'),
+            email: Yup.string().email('Adresă de e-mail invalidă'),
+            consent: Yup.boolean().required(),
+            pickUpDate: Yup.string().required('Este necesar sa selectati o data de ridicare a comenzii dvs.'),
+            pickUpTime: Yup.string().required('Este necesar sa selectati un interval de ridicare a comenzii dvs.'),
+            products: Yup.array().of(
+              Yup.object().shape({
+                quantity: Yup.number().positive().required('Required'),
+                id: Yup.number().positive().required('Required'),
+              })
+            ).min(1, 'The error message if length === 0 | 1'),
+          })}
+          onSubmit={(values, { setSubmitting }) => {
+            const formData = new FormData();
+            formData.append('name', values.name)
+            formData.append('phone', values.phone)
+            formData.append('email', values.email)
+            formData.append('products', JSON.stringify(values.products));
+            formData.append('pickUpDate', values.pickUpDate);
+            formData.append('pickUpTime', values.pickUpTime);
+            axios
+              .post(
+                'https://script.google.com/macros/s/AKfycbx_bdwXRlqbVmbLIy90U5HJG_KwTEjcreTsGJdEC7Bt95bZHzem1K4rFgPwRZES4OYz/exec',
+                formData
+              )
+              .then((response) => {
+                setIsOrderSent(true)
+              })
+              .catch((error) => {
+                console.error('Error submitting form:', error);
+                alert('There was an error submitting the form.');
+              })
+              .finally(() => {
+                setSubmitting(false);
+              });
+          }}
+        >
+          {({
+            values,
+            errors,
+            touched,
+            handleChange,
+            handleBlur,
+            isSubmitting,
+            setFieldValue,
+            isValid
+          }) => (
 
-                          {/* <TextField
+            <>
+              {isOrderSent ? <>
+                <Grid item size={12}>
+                  <Typography variant="h4" align="center" fontWeight="bold">
+                    Formular comandă
+                  </Typography>
+                </Grid>
+                <Grid item size={{ xs: 12, sm: 8, md: 6, lg: 7, xl: 4 }}>
+                  <Form>
+                    <Grid container alignContent={"center"} justifyContent={"center"}>
+                      <Grid item size={{ xs: 12, sm: 12, md: 12, lg: 8 }}>
+                        <TextField
+                          fullWidth
+                          id="name"
+                          name="name"
+                          label="Nume și prenume"
+                          value={values.name}
+                          onChange={handleChange}
+                          onBlur={handleBlur}
+                          error={touched.name && Boolean(errors.name)}
+                          helperText={touched.name && errors.name}
+                          margin="normal"
+                          disabled={isSubmitting}
+                          required
+                        />
+                        <TextField
+                          fullWidth
+                          id="phone"
+                          name="phone"
+                          label="Număr telefon"
+                          value={values.phone}
+                          onChange={handleChange}
+                          onBlur={handleBlur}
+                          error={touched.phone && Boolean(errors.phone)}
+                          helperText={touched.phone && errors.phone}
+                          margin="normal"
+                          disabled={isSubmitting}
+                          required
+                        />
+                        <TextField
+                          select
+                          fullWidth
+                          id="pickupDate"
+                          name="pickupDate"
+                          label="Data ridicare"
+                          value={values.pickupDate}
+                          onChange={(e) => {
+                            const selectedDate = e.target.value;
+                            setFieldValue('pickUpDate', selectedDate);
+                            setPickupIntervals(
+                              selectedDate === "24" ? intervalOptions.restricted : intervalOptions.default
+                            );
+                          }}
+                          margin="normal"
+                          required
+                          disabled={isSubmitting}
+                        >
+                          {dateOptions.map((date) => (
+                            <MenuItem key={date} value={date}>
+                              {`${date} Decembrie`}
+                            </MenuItem>
+                          ))}
+                        </TextField>
+                        <TextField
+                          select
+                          fullWidth
+                          id="pickupInterval"
+                          name="pickupInterval"
+                          label="Interval ridicare"
+                          value={values.pickupInterval}
+                          onChange={(e) => setFieldValue('pickUpTime', e.target.value)}
+                          margin="normal"
+                          required
+                          disabled={!values.pickUpDate || isSubmitting}
+                        >
+                          {pickupIntervals.map((interval, index) => (
+                            <MenuItem key={index} value={interval}>
+                              {interval}
+                            </MenuItem>
+                          ))}
+                        </TextField>
+
+                        {/* <TextField
                         fullWidth
                         id="email"
                         name="email"
@@ -213,40 +214,38 @@ const App = () => {
                         margin="normal"
                         disabled={isSubmitting}
                       /> */}
-                        </Grid>
-                        <Grid item size={{ xs: 12, sm: 12, md: 12, lg: 8 }}>
-                          <AddProduct selectedProductType={selectedProductType} setSelectedProductType={setSelectedProductType} isSubmitting={isSubmitting} products={products} setFieldValue={setFieldValue} values={values} />
-                        </Grid>
-                        <Grid item size={12}>
-                          <OrderList
-                            values={values} handleBlur={handleBlur} handleChange={handleChange} isSubmitting={isSubmitting} errors={errors} touched={touched} setFieldValue={setFieldValue}
-                          />
-                        </Grid>
-                        <Grid item size={12}>
-                          <OrderSubmit
-                            values={values}
-                            isSubmitting={isSubmitting}
-                            isValid={isValid}
-                            handleBlur={handleBlur}
-                            handleChange={handleChange}
-                            errors={errors}
-                            touched={touched}
-                            setFieldValue={setFieldValue}
-                          />
-                        </Grid>
                       </Grid>
-                    </Form>
-                  )}
-                </Formik>
-              </Grid>
-            </>
-            : <Grid item size={12}>
-              <ThankYouMessage setIsOrderSent={setIsOrderSent} />
-            </Grid>}
-
-        </Grid>
-      </div>
-    </div>
+                      <Grid item size={{ xs: 12, sm: 12, md: 12, lg: 8 }}>
+                        <AddProduct selectedProductType={selectedProductType} setSelectedProductType={setSelectedProductType} isSubmitting={isSubmitting} products={products} setFieldValue={setFieldValue} values={values} />
+                      </Grid>
+                      <Grid item size={12}>
+                        <OrderList
+                          values={values} handleBlur={handleBlur} handleChange={handleChange} isSubmitting={isSubmitting} errors={errors} touched={touched} setFieldValue={setFieldValue}
+                        />
+                      </Grid>
+                      <Grid item size={12}>
+                        <OrderSubmit
+                          values={values}
+                          isSubmitting={isSubmitting}
+                          isValid={isValid}
+                          handleBlur={handleBlur}
+                          handleChange={handleChange}
+                          errors={errors}
+                          touched={touched}
+                          setFieldValue={setFieldValue}
+                        />
+                      </Grid>
+                    </Grid>
+                  </Form>
+                </Grid>
+              </>
+                : <Grid item size={12}>
+                  <ThankYouMessage setIsOrderSent={setIsOrderSent} pickUpDate={values.pickUpDate} pickUpTime={values.pickUpTime}/>
+                </Grid>} </>
+          )}
+        </Formik>
+      </Grid>
+    </div >
   );
 };
